isolated functions to its component and removed hooks
