+++ components/ai_chat/resources/page/stories/components_panel.tsx
@@ -47,7 +47,7 @@ export default {
   },
   args: {
     hasQuestions: true,
-    hasSeenAgreement: false,
+    hasSeenAgreement: true,
     currentErrorState: select('Current Status', APIError, APIError.RateLimitReached)
   },
   decorators: [
